[DIDA] 커뮤니티 상세 > 댓글 달았을때 로직 수정
